requirements.txt
tensorflow
streamlit
Pillow
numpy
pandas
streamlit-drawable-canvas
